[B01] Finished sections/:id/pages POST route. Split file controllers into 2. Created appRoute global var.
