+++ server/fileSystem/videoHandler.js
@@ -2,86 +2,12 @@
 
 const config = require('../../config/config.js'),
  	fs = require("fs"),
-	im = require('imagemagick'),
 	ffmpeg = require('fluent-ffmpeg');
 	
-const contentDirectory = __dirname + '/' + config.contentDirectory;
-const imageSizes = config.imageSizes;
+const contentDirectory = global.appRoute + '/' + config.contentDirectory;
 const videoFormats = config.videoFormats;
 
 var convertList = {}; //list of videos being converted atm
-
-function saveImage(image, path) {
-
-	console.log(image);
-
-	if((typeof image) !== 'string') throw `Invalid file, path not equal to string`;
-
-	// save at multiple sizes
-	let promises = [];
-
-	for (var i = imageSizes.length - 1; i >= 0; i--) {
-
-		let width = imageSizes[i];
-		let sizePath = contentDirectory + path.replace('.jpg', '_x' + width + '.jpg');
-
-		promises.push(_resizeAndSaveImage(image, sizePath, width));
-
-	}
-	
-	return Promise.all(promises)
-
-}
-
-function _resizeAndSaveImage(image, path, targetWidth) {
-
-	return new Promise((resolve, reject) => {		
-
-		im.identify(image, function(err, features){
-			if (err) return reject(`File system - Problem identifying image error: ${err}`);
-			
-			let width = (targetWidth < features.width) ? targetWidth : features.width;	//do not make the image larger
-
-			im.resize({
-				srcPath: image,
-				dstPath: path,
-				width:   width,
-				filter: 'Lanczos'
-			}, function(err, stdout, stderr){
-				if (err) {
-					return reject(`File system - Error resizing image size: ${width}, error: ${err}, stdout: ${stdout}, stderr: ${stderr}`)
-				}
-				return resolve();				
-			});
-
-		});	
-	})
-}
-
-function deleteImage(path) {
-
-	let promises = [];
-
-	for (var i = imageSizes.length - 1; i >= 0; i--) {
-		
-		let width = imageSizes[i];
-		let sizePath = contentDirectory + path.replace('.jpg', '_x' + width + '.jpg');
-
-		promises.push(function(callback){
-			
-			return new Promise((resolve, reject) => {	
-
-				fs.unlink(sizePath, function(err) {
-					if(err) return reject(`Error deleting image - path: ${sizePath}, err: ${err}`);
-					return resolve();
-				});
-			})
-		})
-	
-	}
-
-	return promises;
-}
 
 function saveVideo(video, path) {
 
@@ -156,8 +82,5 @@ function deleteVideo(path) {
 	}
 }
 
-
-exports.saveImage = saveImage;
 exports.saveVideo = saveVideo;
-exports.deleteImage = deleteImage;
 exports.deleteVideo = deleteVideo;
